Moved stdout suppressor into utils module
